+++ src/api/chassis/model/threeEncoderSkidSteerModel.cpp
@@ -5,7 +5,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-#include "okapi/impl/chassis/model/threeEncoderSkidSteerModel.hpp"
+#include "okapi/api/chassis/model/threeEncoderSkidSteerModel.hpp"
 
 namespace okapi {
 ThreeEncoderSkidSteerModelArgs::ThreeEncoderSkidSteerModelArgs(
