+++ src/api/chassis/model/xDriveModel.cpp
@@ -1,11 +1,11 @@
-/**
+/**impl
  * @author Ryan Benasutti, WPI
  *
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-#include "okapi/impl/chassis/model/xDriveModel.hpp"
+#include "okapi/api/chassis/model/xDriveModel.hpp"
 #include <utility>
 
 namespace okapi {
